:recycle: [#1471] Remove some code duplication
